+++ common/src/main/java/net/mehvahdjukaar/jepp/rei/JeppReiPlugin.java
@@ -5,13 +5,16 @@ import me.shedaniel.rei.api.client.registry.category.CategoryRegistry;
 import me.shedaniel.rei.api.client.registry.display.DisplayRegistry;
 import me.shedaniel.rei.api.common.category.CategoryIdentifier;
 import me.shedaniel.rei.api.common.util.EntryStacks;
+import me.shedaniel.rei.forge.REIPluginClient;
 import net.mehvahdjukaar.jepp.Jepp;
 import net.mehvahdjukaar.jepp.PaintingInfo;
+import net.minecraft.core.Holder;
 import net.minecraft.core.Registry;
+import net.minecraft.tags.PaintingVariantTags;
 import net.minecraft.world.entity.decoration.PaintingVariant;
 import net.minecraft.world.item.Items;
 
-//@REIPluginClient
+@REIPluginClient
 public class JeppReiPlugin implements REIClientPlugin {
 
     public static final CategoryIdentifier<PaintingInfoDisplay> PAINTING_INFO_TYPE = CategoryIdentifier.of(Jepp.res("painting"));
@@ -24,8 +27,8 @@ public class JeppReiPlugin implements REIClientPlugin {
 
     @Override
     public void registerDisplays(DisplayRegistry registry) {
-        for (PaintingVariant painting : Registry.PAINTING_VARIANT) {
-            PaintingInfo recipe = new PaintingInfoDisplay(painting);
+        for (Holder<PaintingVariant> painting : Registry.PAINTING_VARIANT.getTagOrEmpty(PaintingVariantTags.PLACEABLE)) {
+            PaintingInfo recipe = new PaintingInfoDisplay(painting.value());
             registry.add(recipe);
         }
     }