+++ forge/src/main/java/net/mehvahdjukaar/jepp/forge/JeppForge.java
@@ -1,6 +1,8 @@
 package net.mehvahdjukaar.jepp.forge;
 
+import mezz.jei.forge.JustEnoughItems;
 import net.mehvahdjukaar.jepp.Jepp;
+import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.common.Mod;
 
 /**
@@ -10,6 +12,9 @@ import net.minecraftforge.fml.common.Mod;
 public class JeppForge {
 
     public JeppForge() {
+        if(!ModList.get().isLoaded("jei") && !ModList.get().isLoaded("roughlyenoughitems")){
+            Jepp.LOGGER.error("Jepp requires either JEI or REI mods. None of them was found");
+        }
     }